Enable WarningsAsErrors in csproj files
The changes in this commit include enabling the 'WarningsAsErrors'
setting within the Gnome.Net.Common, Gnome.Net.Glib, and QuickTests
csproj files. With this update, any issue that would normally issue a
warning will now be treated as an error. This should help to minimize
oversights and reinforce strict coding standards.

Signed-off-by: Victor Matia <vmatir@gmail.com>