Make LibraryImportResolver class static
The LibraryImportResolver class has been updated to a static class,
improving the definition of its role. This class now provides a method
for resolving DLL imports by attempting to load a native library with
specified parameters, hence the change to static. This modification
enhances the organization and readability of the code.

Signed-off-by: Victor Matia <vmatir@gmail.com>