Remove unused Marshalling imports and update DateTimeException comments
Unused Marshalling import statements were removed from multiple files
in the Gnome.Net.Glib project, making the code cleaner and less
cluttered.

Additionally, the comment documentation for DateTimeException has been
updated to provide a more comprehensive understanding of the class's
purpose and functionality.

A small change was also made in the DateTime class to remove the
nullable check for the 'other' object in the Equals and CompareTo
methods.

Signed-off-by: Victor Matia <vmatir@gmail.com>